src: move sources into sources/
Redmine issue:
https://redmine.openinfosecfoundation.org/issues/3166
