+++ src/runmodes.c
@@ -48,7 +48,7 @@
 
 #include "log-httplog.h"
 
-#include "source-pfring.h"
+#include "sources/source-pfring.h"
 
 #include "tmqh-flow.h"
 #include "flow-manager.h"
