+++ src/stream-tcp.c
@@ -75,7 +75,7 @@
 #include "util/util-runmodes.h"
 #include "util/util-random.h"
 
-#include "source-pcap-file.h"
+#include "sources/source-pcap-file.h"
 
 //#define DEBUG
