+++ src/win32/win32-misc.c
@@ -28,7 +28,7 @@
 #include "suricata-common.h"
 #include "win32-misc.h"
 #include "direct.h"
-#include "util-ip.h"
+#include "util/util-ip.h"
 
 void setenv(const char *name, const char *value, int overwrite)
 {
